2c

+++ src/model/Product.java
@@ -3,10 +3,10 @@ package model;
 public class Product {
 
     private String name;
-    private int price;
+    private double price;
     private String universe;
 
-    public Product(String name, int price, String universe) {
+    public Product(String name, double price, String universe) {
         this.name = name;
         this.price = price;
         this.universe = universe;
@@ -20,11 +20,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
